refactor(memory): extract query embedding breaker
RetrievalService carried the warm-recall query embedding circuit as two
private maps plus eight private methods next to the recall pipeline. Move
that state machine into QueryEmbeddingCircuitBreaker so recall only asks
for a started promise, a circuit state, or a reset, and the breaker no
longer leaks its in-flight entry to callers.
